requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flask-wtf==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
cryptography==41.0.5

